% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{adolphi_tiepoint_simmer}
\alias{adolphi_tiepoint_simmer}
\title{Simulate from Adolphi tie-point distributions
Returns the probability distributions for the tie-points given by Adolphi et al. (2018) and Muscheler et al. (2020)}
\usage{
adolphi_tiepoint_simmer(
  nsims = 10000,
  tieshifts = numeric(5),
  plotdens = FALSE,
  x.ref = NULL,
  plothist = list(compute = FALSE, breaks = 50, col = "orange")
)
}
\arguments{
\item{nsims}{Integer. The number of samples to be produced.}

\item{tieshifts}{numeric which gives the amount each tie-point should be shifted, e.g. to go from describing offset from GICC05 to years before present.}

\item{plotdens}{boolean that is passed on to adolphiloader. Set \code{TRUE} if you want to plot distributions}

\item{x.ref}{numeric. Gives reference value on the x-axis when distributions are plotted.}

\item{plothist}{list object describing if and how histogram of samples should be plotted}
}
\value{
returns a list containing the pdfs for each tie-point.
}
\description{
Simulate from Adolphi tie-point distributions
Returns the probability distributions for the tie-points given by Adolphi et al. (2018) and Muscheler et al. (2020)
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{adolphi}
\keyword{pdf}
\keyword{tiepoint}
