% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{adolphiloader}
\alias{adolphiloader}
\title{Load Adolphi tie-point distributions
Returns the probability distributions for the tie-points given by Adolphi et al. (2018) and Muscheler et al. (2020)}
\usage{
adolphiloader(tieshifts = numeric(5), plotdens = FALSE, x.ref = NULL)
}
\arguments{
\item{tieshifts}{numeric which gives the amount each tie-point should be shifted, e.g. to go from describing offset from GICC05 to years before present.}

\item{plotdens}{boolean. If \code{TRUE} plot pdfs.}

\item{x.ref}{numeric. Gives reference value on the x-axis when distributions are plotted.}
}
\value{
returns a list containing the pdfs for each tie-point.
}
\description{
Load Adolphi tie-point distributions
Returns the probability distributions for the tie-points given by Adolphi et al. (2018) and Muscheler et al. (2020)
}
\references{
Adolphi, F., Bronk Ramsey, C., Erhardt, T., Edwards, R. L., Cheng, H., Turney, C. S. M., Cooper, A., Svensson, A., Rasmussen, S. O., Fischer, H., and Muscheler, R. (2018).
Connecting the Greenland ice-core and U∕Th timescales via cosmogenic radionuclides: testing the synchroneity of Dansgaard–Oeschger events,
Clim. Past, 14, 1755–1781, https://doi.org/10.5194/cp-14-1755-2018

Muscheler, R., Adolphi, F., Heaton, T., Bronk Ramsey, C., Svensson, A., Van der Plicht, J., & Reimer, P. (2020).
Testing and Improving the IntCal20 Calibration Curve with Independent Records.
Radiocarbon, 62(4), 1079-1094. doi:10.1017/RDC.2020.54
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{adolphi}
\keyword{simulation}
\keyword{tiepoint}
