% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla.R
\name{bremla}
\alias{bremla}
\title{Bayesian regression modeling of layered archives}
\usage{
bremla(
  formula,
  data,
  reference.label = NULL,
  nsims = 10000,
  events = NULL,
  synchronization = NULL,
  control.fit = NULL,
  control.sim = NULL,
  control.linramp = NULL,
  control.transition_dating = NULL,
  control.bias = NULL,
  print.progress = FALSE
)
}
\arguments{
\item{formula}{formula describing the predictor used in fitting the data. Partial covariates (psi)
can be filled in by specifying the degree and events in \code{events} argument.}

\item{data}{data.frame containing response and covariates. Must include 'age' and 'depth'.
The first row is used for initializing 'age' and 'depth', hence the rest can be set to NA.
Covariates related to the events (psi) can be filled in automatically by specifying the degree and events in \code{events} argument.}

\item{reference.label}{Character label of reference timescale. Used in \code{\link{plot},\link{summary}}.}

\item{nsims}{Number of chronologies to be simulated.}

\item{events}{List object describing locations, degree and other informations about
the specific climatic periods used in the predictor. If used must include at least
\code{events\$locations}. See \code{\link{events.default}} for more details.}

\item{synchronization}{List containing locations and details for sampling tie-points used to synchronize our time scale.}

\item{control.fit}{List containing specifications for the fitting procedure.
See \code{\link{control.fit.default}} for more details.}

\item{control.sim}{List containing specifications for sampling chronologies (synchronous or otherwise).
See \code{\link{control.sim.default}} for more details.}

\item{control.linramp}{List containing specifications for fitting a linear ramp model to given data.
If used, must include \code{control.linramp\$proxy} and \code{control.linramp\$interval}. See
\code{\link{control.linramp.default}} for details.}

\item{control.transition_dating}{List containing specifications for
estimating a given transition. See \code{\link{control.transition_dating.default}} for more details.}

\item{control.bias}{List containing specifications for adding a potential unknown stochastic counting bias.
See \code{\link{control.bias.default}} for more details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}
}
\value{
Returns an S3 object of class "bremla". This includes output from all functions nested within the bremla function. Including fitted marginals and summary statistics, simulated chronologies, time spent on each step.
}
\description{
Fits a regression model to the data and produces chronology simulations.
}
\seealso{
\code{\link{bremla_modelfitter},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
