% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_biased_chronologies.R
\name{bremla_biased_chronologies}
\alias{bremla_biased_chronologies}
\title{Linear ramp fit}
\usage{
bremla_biased_chronologies(object, control.bias, print.progress = FALSE)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{bremla_chronology_simulation}}}

\item{control.bias}{List containing specifications for how the unknown random bias
should be implemented. See \code{\link{control.bias.default}} for details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to the screen.}
}
\value{
Returns the same \code{object} list from the input, but appends another
list of the summary statistics for each analysis (and samples if \code{store.samples=TRUE}),
and inputs (\code{settings}). These are stored in \code{object\$biases.}
}
\description{
Fits the linear ramp described by Erhardt et al. (2019) to proxy values using INLA.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")
age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O
eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])
nsims=5000
object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=nsims)
object = bremla_modelfitter(object)
object = bremla_chronology_simulation(object,nsims=nsims)
object = bremla_biased_chronologies(object,bias.model="uniform",nsims=nsims,
           biasparams=cbind( c(1,1),c(0.98,1.02),c(0.96,1.04) ) )
plot(object)
}
}
\seealso{
\code{\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bias}
\keyword{bremla}
