% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_modelfitter.R
\name{bremla_modelfitter}
\alias{bremla_modelfitter}
\title{Bremla model fitting}
\usage{
bremla_modelfitter(object, control.fit, print.progress = FALSE)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{bremla_prepare}}}

\item{control.fit}{List containing specifications for fitting procedure. See
\code{\link{control.fit.default}} for details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}
}
\value{
Returns the same \code{object} list from the input, but appends information
from the fitting procedure in \code{object\$fitting}. Including fitted values, residuals, posterior distributions
and summary statistics for both fixed and random effects.
}
\description{
Fits the bremla linear regression model to observations.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])

object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=0)
object = bremla_modelfitter(object)
summary(object)
}


}
\seealso{
\code{\link{bremla_prepare},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{fitting}
