% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_prepare.R
\name{bremla_prepare}
\alias{bremla_prepare}
\title{Bremla preparation function}
\usage{
bremla_prepare(
  formula,
  data,
  reference.label = NULL,
  nsims = NULL,
  events = NULL,
  synchronization = NULL,
  control.fit = NULL,
  control.sim = NULL,
  control.linramp = NULL,
  control.transition_dating = NULL,
  control.bias = NULL
)
}
\arguments{
\item{formula}{Formula describing the response and covariates (partial
covariates (phi) can be filled out using \code{events$fill_formula})}

\item{data}{data.frame including response, all covariates included in
\code{formula}, as well as 'depth' and 'age' describing both axes of reference chronology.
First row is used to extract initial values 'y0' and 'z0', the remaining variables
in the first row is typically not used and can be left as \code{NA}. Partial
covariates (phi) can be filled out using 'events$fill_data'}

\item{reference.label}{Character label of reference timescale. Used in \code{\link{plot},\link{summary}}.}

\item{nsims}{Number of chronologies to be simulated.}

\item{events}{List object describing the specifics of the climate transitions
used in 'formula'. Must include an item called \code{locations}. See \code{?events.default} for details.}

\item{synchronization}{List object describing specifics related to tie-points
and their distribution. If \code{synchronization$method="adolphi"} a specific
set of instructions are employed. If not, this list must include \code{locations}
argument and preferably \code{locations_unit} to determine which axis (default \code{unit="depth"}).
See \code{\link{synchronization.default}} for details.}

\item{control.fit}{List object describing specifics related to the fitting procedure.
See \code{\link{control.fit.default}} for details.}

\item{control.sim}{List object describing specifics related to simulating chronologies.
See \code{\link{control.sim.default}} for details.}

\item{control.linramp}{List containing specifications for fitting a linear ramp model to given data.
If used, must include \code{control.linramp\$proxy} and \code{control.linramp\$interval}. See
\code{\link{control.linramp.default}} for details.}

\item{control.transition_dating}{List object describing specifics related to
the estimation of a given transition. Must include \code{interval} and \code{proxy}
used in linear ramp estimation. See \code{\link{control.transition_dating.default}} for details.}

\item{control.bias}{List object describing specifics related to the simulation of
unknown stochastic bias. See \code{\link{control.bias.default}} for details.}
}
\value{
Returns a list containing data in internal formatting (\code{data}), linear regression formula (\code{ls.formula}), inla formula (\code{formula}), input settings and indices corresponding to climate transitions (\code{.args}) and z_0, y_0 and x_0 (\code{preceeding})
}
\description{
Prepares input and formats data for bremla and functions therein.
}
\examples{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])

object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=0)

}
\seealso{
\code{\link{bremla},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{preparation}
