% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{bremla_simulationsummarizer}
\alias{bremla_simulationsummarizer}
\title{Simulation-summarizer}
\usage{
bremla_simulationsummarizer(
  object,
  CI.type = "hpd",
  sync = TRUE,
  print.progress = FALSE
)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{bremla_chronology_simulation}}}

\item{CI.type}{character describing which uncertainty intervals should be used (\code{"quantiles" or "hpd"}). Highest posterior density is used by default}

\item{sync}{boolean. Set \code{TRUE} if summary statistics should be computed for syncrhonized samples (if any). \code{FALSE} for unsynchronized.}

\item{print.progress}{Boolean describing whether or not progress should be printed on screen.}
}
\value{
Returns the \code{object} list from the input and appends additional summary statistics: posterior marginal mean and uncertainty interval-
}
\description{
Computes posterior marginal mean and uncertainty intervals from simulations.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])

object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=0)
object = bremla_modelfitter(object)
object = bremla_chronology_simulation(object,nsims=1000)
object = bremla_simulationsummarizer(object,CI.type="quant",sync=FALSE)
plot(object)
}
}
\seealso{
\code{\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{indices}
