% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buizert_onsets.R
\docType{data}
\name{buizert_onsets}
\alias{buizert_onsets}
\title{Buizert transition onsets}
\format{
A \code{tbl} \code{data.frame} with 22 rows and 5 columns:
\describe{
  \item{Transition}{Name of the transition.}
  \item{Depth}{NGRIP depth of the onset of the transition (in meters).}
  \item{Age}{NGRIP depth of the onset of the transition (in years before present).}
  \item{`Age uncertainty`}{The dating uncertainty of the transition onset (in years).}
  \item{Midpoint}{The midpoint of the transition (in years).}
  }
}
\usage{
buizert_onsets
}
\description{
Transition onsets of abrupt warmings in the northern hemisphere. Taken from
Table 1 of Buizert et al. (2015). This data set is only used to compare
our transition dating estimation in "Examples/Paper_results/Paper_results_1.Rmd"
and "Examples/Paper_results/Paper_results_2.Rmd".
}
\references{
Buizert, C., Cuffey, K. M., Severinghaus, J. P., Baggenstos, D.,
Fudge, T. J., Steig, E. J., Markle, B. R., Winstrup, M., Rhodes, R. H.,
Brook, E. J., Sowers, T. A., Clow, G. D., Cheng, H., Edwards, R. L.,
Sigl, M., McConnell, J. R., and Taylor, K. C.: The WAIS Divide deep ice core
WD2014 chronology – Part 1: Methane synchronization (68–31 ka BP) and the gas
age–ice age difference, Clim. Past, 11, 153–173, https://doi.org/10.5194/cp-11-153-2015,
2015.
}
\keyword{datasets}
