% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capron_onsets_NGRIP_d18O.R
\docType{data}
\name{capron_onsets_NGRIP_d18O}
\alias{capron_onsets_NGRIP_d18O}
\title{Capron transition onsets (d18O)}
\format{
A \code{tbl data.frame} with 25 rows and 16 columns. The following columns
are relevant to analyses carried out in this package:
\describe{
  \item{`NGRIP d18O`}{Name of the transition.}
  \item{`t1 (50\%)`}{The median of the transition onset in years before present
  (used as point estimate).}
  \item{`d1 (50\%)`}{The median of the transition onset depth in meters.}
  }
}
\usage{
capron_onsets_NGRIP_d18O
}
\description{
Abrupt warmings observed in the d18O record of the NGRIP ice core. Taken from
Supplementary data 4 in Capron et al. (2021). This data set is only used to compare
our transition dating estimation in "Examples/Paper_results/Paper_results_1.Rmd"
and "Examples/Paper_results/Paper_results_2.Rmd".
}
\references{
Capron, E., Rasmussen, S.O., Popp, T.J. et al. The anatomy of past
abrupt warmings recorded in Greenland ice. Nat Commun 12, 2106 (2021).
https://doi.org/10.1038/s41467-021-22241-w
}
\keyword{datasets}
