% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_arguments.R
\name{control.bias.default}
\alias{control.bias.default}
\title{Default variables in control.bias}
\usage{
control.bias.default()
}
\value{
Returns a list including default values for all variables in \code{control.bias}.
}
\description{
Sets the default variables in the list \code{control.bias} used to specify the
how an unknown stochastic bias should be included. The list contains the following arguments:
\describe{
  \item{\code{bias.model} }{Character describing which model should be used for the unknown bias.
  Currently only \code{"uniform"} is supported.}
  \item{\code{biasparams} }{Gives the parameters for the \code{bias.model}. If \code{ncol}>1,
  the function will be applied to each column consecutively.}
  \item{\code{nsims} }{Integer giving the number of samples to be produced for the model. Default value
  is \code{10000}.}
  \item{\code{store.samples} }{Boolean indicating if samples should be stored. Default value is \code{FALSE}.}
}
}
\seealso{
\code{\link{bremla}, \link{bremla_biased_chronologies},
\link{set.options}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bias}
\keyword{bremla}
\keyword{default}
