% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_arguments.R
\name{control.transition_dating.default}
\alias{control.transition_dating.default}
\title{Default variables in control.transition_dating}
\usage{
control.transition_dating.default()
}
\value{
Returns a list including default values for all variables in
\code{control.transition_dating}.
}
\description{
Sets the default variables in the list \code{control.transition_dating} used to
specify the procedure for estimating the dating uncertainty of a given transition
following a linear ramp model fit.
The list contains the following arguments:
   \itemize{
     \item \code{label} Character used to describe the name of the transition in future \code{plot}. Default \code{NULL}.
     \item \code{nsims} Integer giving the number of samples to be generated for onset transition age. Default value is \code{10000}.
     \item \code{sync} Boolean indicating if the synchronized age-depth model should be used. Default value is \code{TRUE}, but this is set to \code{fALSE} if synchronized samples are not yet computed.
     \item \code{age.label} Character describing what label to place on x-axis (age) in a future \code{plot}. Default value is \code{""}.
     \item \code{age.ref} Numeric giving a reference value for the age which will be included in a future \code{plot}.
     }
}
\seealso{
\code{\link{bremla}, \link{linrampfitter},
\link{events_depth_to_age},\link{set.options}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{default}
\keyword{onset}
\keyword{transition}
