% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DO_depth_to_age.R
\name{DO_depth_to_age}
\alias{DO_depth_to_age}
\title{Complete dating uncertainty of DO-onsets}
\usage{
DO_depth_to_age(
  object,
  nsims = 10000,
  print.progress = FALSE,
  label = NULL,
  age.reference = NULL
)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{linrampfitter}}}

\item{nsims}{Integer denoting the number of simulations of DO onset age. Cannot exceed the number of simulated age chronologies from \code{\link{bremla_chronology_simulation}}.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}

\item{label}{character string describing the label designed the transition, e.g. \code{label="GI-11"}.}

\item{age.reference}{numeric denoting a reference value such as the transition onset age obtained by other means. This appears when using \code{plot}. If \code{depth.reference=NULL} then this is not used.}
}
\value{
Returns the same \code{object} list from the input, but appends the simulated DO onset ages along with summary statistics.
}
\description{
Combines linear ramp model fit and Bayesian regression modeling to estimate complete dating uncertainty of the onset of Dansgaard-Oeschger events using Monte Carlo simulation.
}
\examples{

}
\seealso{
\code{\link{bremla_chronology_simulation},\link{linrampfitter}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{DO}
\keyword{bremla}
\keyword{dating}
