% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_arguments.R
\name{events.default}
\alias{events.default}
\title{Default variables in events}
\usage{
events.default()
}
\value{
Returns a list including default values for all variables in \code{events}.
}
\description{
Sets the default variables in the list \code{events} used to specify the climatic
periods and separating events used in the linear predictor. The list contains the following arguments:
\itemize{
  \item{\code{locations} }{Numeric describing the location of the transitions between climatic periods.
   Default value is \code{NULL}, and must be provided by user.}
  \item{\code{locations_unit} }{Character describing on which axis the locations are given. Can be \code{age},
  \code{depth} or \code{index}. Default value is \code{depth}.}
  \item{\code{degree} }{Integer giving the polynomial degree the individual period-specific
  covariates should be in. Set to \code{0} to only include constants between transitions,
  \code{1} to also include a linear function between covariates and \code{2} to include a square
  function. Default value is \code{1}.}
  \item{\code{fill_data} }{Boolean telling whether or not the \code{data} object should be filled to include
  the climate period covariates. Default value is \code{TRUE}.}
  \item{\code{fill_formula} }{Boolean telling whether or not the \code{formula} object should be
  filled to include description of the period-specific covariates (with degree
  as given in to \code{events\$degree}.). Default value is \code{TRUE}.}
}
}
\seealso{
\code{\link{bremla},\link{set.options}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{default}
\keyword{events}
