% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISevents.R
\docType{data}
\name{GISevents}
\alias{GISevents}
\title{Climatic transitions in NGRIP}
\format{
A data.frame with 117 rows and 5 columns:
\describe{
  \item{Event}{Description of event}
  \item{NGRIP.depth..m.}{Recorded depth in NGRIP (meter)}
  \item{Age..a.b2k}{Corresponding GICC05 age in years before 2000 AD}
  \item{Event.definition.uncertainty}{The dating uncertainty arising from the layer counting process.}
  \item{Maximum.counting.error..years.}{The maximum counting error of each climate event (half the number of uncertain layers).}
  }
}
\source{
\url{}
}
\usage{
GISevents
}
\description{
A data set containing the onsets of climatic events as recorded in the NGRIP ice core records.
}
\references{
Will be added soon!
}
\keyword{datasets}
