% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{inlaloader}
\alias{inlaloader}
\title{Load INLA safely
Wrapper function that loads the INLA package for use in tests.}
\usage{
inlaloader()
}
\value{
Boolean. \code{TRUE} if INLA was loaded successfully, \code{FALSE} otherwise.
}
\description{
Load INLA safely
Wrapper function that loads the INLA package for use in tests.
}
\examples{
\dontrun{
if(inlaloader()){
    #Code that depend on INLA
  }
}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
