% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{linramprev}
\alias{linramprev}
\title{Linear ramp function (reverse)}
\usage{
linramprev(t, t0 = 0, dt = 1, y0 = 0, dy = 1)
}
\arguments{
\item{t}{Vector describing the x-axis (here, depth)}

\item{t0}{the onset of the transition}

\item{dt}{The transition duration}

\item{y0}{The initial level of the observed values (here, proxy values)}

\item{dy}{The increase in the observed values}
}
\value{
Returns a vector y-axis corresponding to \code{t}
}
\description{
Computes the (noiseless) linear ramp function, but in reverse.
}
\seealso{
\code{\link{linrampfitter}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{linramp}
