% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGRIP5cm_d18O_GICC05.R
\docType{data}
\name{NGRIP5cm_d18O_GICC05}
\alias{NGRIP5cm_d18O_GICC05}
\title{NGRIP d18O and dust records (sheet 2)}
\format{
A data.frame with 18672 rows and 4 columns:
\describe{
  \item{NGRIP_depth.m}{NGRIP depth in equidistant 5 cm intervals.}
  \item{delta_18O.permil}{Measured del-18O isotope (permil).}
  \item{GICC05_age.yb2k}{Corresponding GICC05 age (y b2k).}
  \item{GICC05_MCE.yr}{The GICC05 maximum counting error (yr).}
  }
}
\source{
\url{}
}
\usage{
NGRIP5cm_d18O_GICC05
}
\description{
The second sheet (corresponding to NGRIP-2) of the NGRIP d18O and dust records.
}
\references{
Will be added soon!
}
\keyword{datasets}
