% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bremla.R
\name{plot.bremla}
\alias{plot.bremla}
\title{Plot bremla model}
\usage{
\method{plot}{bremla}(
  x,
  plot.ls = list(fitted = TRUE, legend = NULL, residuals = TRUE, histogram = TRUE, qqplot
    = FALSE, acf = TRUE, xrev = FALSE, label.fit = NULL, label.res = NULL, label.hist =
    NULL, label.qq = NULL, label.acf = NULL, merge = TRUE),
  plot.inla.posterior = list(posteriors = TRUE, label = NULL),
  plot.inlasims = list(plotsims = 0, legend = NULL, xrev = FALSE, label = NULL),
  plot.syncsims = list(plotsims = 0, legend = NULL, xrev = FALSE, label = NULL),
  plot.tiepoints = list(col.hist = "orange", breaks.hist = 50, label.x = "Age (yb2k)",
    label.main = NULL),
  plot.bias = list(MCE = NULL, legend = NULL, xrev = FALSE, label = NULL),
  plot.linramp = list(depth.reference = NULL, show.t0 = TRUE, show.t1 = TRUE, xrev =
    TRUE, label = NULL),
  plot.event_depth = list(depth.reference = NULL, xrev = TRUE, label = NULL),
  plot.event_age = list(age.reference = NULL, xrev = TRUE, label = NULL),
  postscript = FALSE,
  pdf = FALSE,
  prefix = "bremla.plots/figure-",
  ...
)
}
\arguments{
\item{x}{\code{bremla} S3 class. Output of \code{\link{bremla}} function}

\item{plot.ls}{List specifying how the least square fit should be illustrated. \code{fitted=TRUE} plots the fitted values with label \code{label.fit}, \code{legend} specifies the legend, \code{residuals=TRUE} means the residuals are plotted with title \code{label.res}.
if \code{histogram=TRUE} the residuals are represented in a histogram with label \code{label.hist}, if \code{qqplot=TRUE} a quantile-quantile plot of the residuals are plotted with title \code{label.qq} and if
\code{acf=TRUE} the empirical autocorrelations are plotted with title \code{label.acf}.}

\item{plot.inla.posterior}{list specifying how the results from the inla regression fit should be plotted. If \code{posteriors=TRUE} then the posterior marginal distributions of the hyperparameters are plotted with title \code{label}.}

\item{plot.inlasims}{list specifying how the simulated chronologies from the INLA posterior should be plotted. \code{plotsims} gives how many simulated chronologies should be included in the plot (with title \code{label}), \code{legend} specifies the legend, if \code{xrev=TRUE} the x-axis is reversed to give a chronological ordering.}

\item{plot.syncsims}{list specifying how the simulated synchronized chronologies from the INLA posterior should be plotted. \code{plotsims} gives how many simulated chronologies should be included in the plot (with title \code{label}), \code{legend} specifies the legend, if \code{xrev=TRUE} the x-axis is reversed to give a chronological ordering.}

\item{plot.tiepoints}{list containing specifications for plotting histogram of sampled tie-points.}

\item{plot.bias}{list specifying how the simulations under the assumptions of unknown counting bias should be represented. If \code{MCE} is given as a numeric vector it will be included in the plot (with title \code{label}). \code{legend} specifies the legend, if \code{xrev=TRUE} the x-axis is reversed to give a chronological ordering.}

\item{plot.linramp}{list specifying how the linear ramp fit should be plotted. If \code{depth.reference} is given, it will be represented by a vertical dotted line. If \code{show.t0=TRUE} the posterior marginal distribution of the onset is included (non-normalized). If \code{show.t1=TRUE} the posterior marginal distribution of the end point of the transition is included (non-normalized). If \code{xrev=TRUE} the x-axis is reversed to give a chronological ordering. \code{label} gives the title of the plot.}

\item{plot.event_depth}{list specifying how the posterior distribution of the onset depth should be plotted. If \code{depth.reference} is given, it will be represented by a vertical dotted line. If \code{xrev=TRUE} the x-axis is reversed to give a chronological ordering. \code{label} gives the title of the plot.}

\item{plot.event_age}{list specifying how the histogram of the simulated onset ages should be plotted. If \code{age.reference} is given, it will be represented by a vertical dotted line. If \code{xrev=TRUE} the x-axis is reversed to give a chronological ordering. \code{label} gives the title of the plot.}

\item{postscript}{Boolean variable indicating if postscript files should be produced instead.}

\item{pdf}{Boolean variable indicating if pdf files should be produced instead.}

\item{prefix}{The prefix for created files. Additional numbering is added.}

\item{...}{Additional arguments to \code{postscripts()}, \code{pdf()} or \code{dev.new()}.}
}
\description{
Plots results from bremla S3 class.
}
\examples{
\donttest{
if(inlaloader()){
require(stats)
set.seed(1)
n <- 1000
phi <- 0.8
sigma <- 1.2
a_lintrend <- 0.3; a_proxy = 0.8
dy_noise <- as.numeric(arima.sim(model=list(ar=c(phi)),n=n,sd=sqrt(1-phi^2)))
lintrend <- seq(from=10,to=15,length.out=n)

proxy <- as.numeric(arima.sim(model=list(ar=c(0.9)),n=n,sd=sqrt(1-0.9^2)))
dy <- a_lintrend*lintrend + a_proxy*proxy + sigma*dy_noise

y0 = 11700;z0=1200
age = y0+cumsum(dy)
depth = 1200 + 1:n*0.05


formula = dy~-1+depth2 + proxy
data = data.frame(age=age,dy=dy,proxy=proxy,depth=depth,depth2=depth^2)
data = rbind(c(y0,NA,NA,z0,NA),data) #First row is only used to extract y0 and z0.

events=list(locations=c(1210,1220,1240))
control.fit = list(ncores=2,noise="ar1")
control.sim=list(synchronized=2,
                 summary=list(compute=TRUE))

object = bremla_prepare(formula,data,nsims=5000,reference.label="simulated timescale",
                        events = events,
                        control.fit=control.fit,
                        control.sim=control.sim)
object = bremla_modelfitter(object)
object = bremla_chronology_simulation(object, print.progress=TRUE)
summary(object)
plot(object)
}
}

}
\seealso{
\code{\link{bremla},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{plot}
