% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bremla.R
\name{print.bremla}
\alias{print.bremla}
\title{Print bremla model}
\usage{
\method{print}{bremla}(x, digits = 4L, ...)
}
\arguments{
\item{x}{\code{bremla} S3 class. Output of \code{\link{bremla}} function}

\item{digits}{Number of digits displayed.}

\item{...}{Other arguments}
}
\description{
Prints essentials from bremla S3 class.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
object = bremla(age,depth,proxy,events=eventdepths,nsims=100,
  synchronization = list(locations=c(11050,12050,13050,22050,42050),
                          locations.type="age",method="adolphi",
                          samples=NULL),
  print.progress=TRUE
  )
print(object)
}
}
\seealso{
\code{\link{bremla},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{sumary}
