% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bremla.R
\name{print.summary.bremla}
\alias{print.summary.bremla}
\title{Print summary.bremla class}
\usage{
\method{print}{summary.bremla}(x, digits = 4L, ...)
}
\arguments{
\item{x}{\code{summary.bremla} S3 class. Output of \code{\link{summary.bremla}} function}

\item{digits}{Number of digits displayed.}

\item{...}{Other arguments}
}
\description{
Prints summary.bremla S3 class.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
object = bremla(age,depth,proxy,events=eventdepths,nsims=100,
  synchronization = list(locations=c(11050,12050,13050,22050,42050),
                          locations.type="age",method="adolphi",
                          samples=NULL),
  print.progress=TRUE
  )
ss = summary(object)
print(ss)
}

}
\seealso{
\code{\link{bremla},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{summary.print}
