% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{Qmaker_ar1cum}
\alias{Qmaker_ar1cum}
\title{Precision matrix of cumulative AR(1) process
Produces the sparse precision matrix of a cumulative AR(1) process}
\usage{
Qmaker_ar1cum(n, sigma, rho)
}
\arguments{
\item{n}{The dimension of the multivariate process.}

\item{sigma}{The standard deviation (scaling parameter).}

\item{rho}{The lag-one correlation coefficient of the AR(1) process.}
}
\value{
Returns sparse precision matrix of dimension \code{n x n}.
}
\description{
Precision matrix of cumulative AR(1) process
Produces the sparse precision matrix of a cumulative AR(1) process
}
\examples{
n=100
sigma=1
rho=0.8
Q = Qmaker_ar1cum(n,sigma,rho)

nsims = 1000
muvek = seq(1,5,length.out=n)
samples = Qsimmer(nsims,Q,muvek)

}
\seealso{
\code{\link{Qsimmer}, \link{Qymaker}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{GMRF}
\keyword{matrix}
\keyword{precision}
\keyword{sparse}
