% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeneric.uneven.AR1.R
\name{rgeneric.uneven.AR1}
\alias{rgeneric.uneven.AR1}
\title{rgeneric model specification of linear ramp function}
\usage{
rgeneric.uneven.AR1(
  cmd = c("graph", "Q", "mu", "initial", "log.norm.const", "log.prior", "quit"),
  theta = NULL
)
}
\arguments{
\item{cmd}{Vector containing list of function names necessary for the rgeneric model.}

\item{theta}{Vector describing the hyperparameters in internal scaling.}
}
\value{

}
\description{
Non-standard models in INLA requires specification using the rgeneric modeling framework.
This includes key functions that provide information on precision matrix, mean vector, priors, graph etc.
This is intended for internal use only, but the documentation is included here in case someone want to change something.
 See example below for how rgeneric is used can be used.
}
\examples{

}
\seealso{
\code{\link{inla},\link{linrampfitter}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{inla}
\keyword{rgeneric}
