% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{skewsampler}
\alias{skewsampler}
\title{Skew-quasi-Gaussian sampler
Produces simulations from the merged Gaussian process}
\usage{
skewsampler(
  n,
  mode = 0,
  sdL = 1,
  sdU = 1,
  log = FALSE,
  plothist = list(compute = FALSE, breaks = 50, col = "orange")
)
}
\arguments{
\item{n}{The number of samples to produce.}

\item{mode}{Mode where the two Gaussians are merged.}

\item{sdL}{lower standard deviation.}

\item{sdU}{upper standard deviation.}

\item{log}{boolean describing whether or not the log value should be returned.}

\item{plothist}{list object with arguments describing how (and if) a histogram should be produced.}
}
\value{
returns a numeric object with \code{n} samples.
}
\description{
Skew-quasi-Gaussian sampler
Produces simulations from the merged Gaussian process
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{sample}
\keyword{skew}
\keyword{tiepoint}
