% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{which.index}
\alias{which.index}
\title{Corresponding index}
\usage{
which.index(events, record, suppress = TRUE)
}
\arguments{
\item{events}{Vector describing the values of interest (here: either depth or age of events)}

\item{record}{Vector we want to search (here: either full depth or age record)}

\item{suppress}{Boolean. Set to \code{TRUE} to suppress warnings that events fall outside designated interval.}
}
\value{
Returns a vector of indices for which elements of \code{record} best matches the elements of \code{events}-
}
\description{
Finds the indices where events best match values in a given vector.
}
\seealso{
\code{\link{bremla}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{indices}
