% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla.R
\name{bremla}
\alias{bremla}
\title{Bayesian regression modeling of layered archives}
\usage{
bremla(
  age,
  depth,
  proxy,
  events = NULL,
  nsims = 10000,
  eventmeasure = "depth",
  reference.label = NULL,
  proxy.type = "d18O",
  transform = "identity",
  reg.model = list(const = FALSE, depth1 = FALSE, depth2 = TRUE, proxy = TRUE, psi0 =
    TRUE, psi1 = TRUE),
  noise = "ar1",
  method = "inla",
  CI.type = "quantiles",
  synchronization = NULL,
  event.estimation = NULL,
  bias = NULL,
  store.everything = FALSE,
  print.progress = FALSE
)
}
\arguments{
\item{age}{Vector of observed ages y_0,...,y_n}

\item{depth}{Vector of depths z_0,...,z_n corresponding to \code{age}}

\item{proxy}{Vector of observed proxy (e.g. del-18O) at depths \code{depth}}

\item{events}{Vector describing locations (in dimension specified by \code{eventmeasure}) of climate transitions for use in linear regression model}

\item{nsims}{Number of chronologies to be simulated}

\item{eventmeasure}{Character describing in which dimension the climate events are located ("age", "depth", "index")}

\item{reference.label}{String denoting the label used for age data/reference}

\item{proxy.type}{String denoting which proxy is used (for plot-labels).}

\item{transform}{String denoting which transformation should be applied to the residuals. Default "identity", "log" is also supported, but experimental.}

\item{reg.model}{List of booleans specifies which effects to be included in the linear regression model: intercept (\code{const}), linear response to depth (\code{depth1}), quadratic response to depth (\code{depth2}), linear response to proxy (\code{proxy}), individual intercepts for each climate period (\code{psi0}), individual linear responses to depth for each climate period (\code{psi1})}

\item{noise}{Character specifying the noise model: independent identically distributed (\code{iid}), first order autoregressive (\code{ar1}) or second order autoregressive (\code{ar2})}

\item{method}{Character specifying how the model is fitted. Currently only least squares (\code{ls}) and INLA (\code{inla}) are supported, and least squares is always run.}

\item{CI.type}{Character describing which type of uncertainty intervals to be used. \code{CI.type="quantiles"} is computed from the quantiles, \code{CI.type="hpd"} computes highest posterior density intervals (takes a little longer to compute). If distribution is unimodal and symmetric these are equivalent.}

\item{synchronization}{List containing information used for simulating tie-points and synchronizing age-depth model to them.}

\item{event.estimation}{List specifying the settings for the linear ramp fit and subsequent age simulation of a specified transition. See example.}

\item{bias}{List specifying the settings for the simulation of chronologies subjected to unknown counting bias. See example for implementation.}

\item{store.everything}{Boolean describing whether optional information (e.g. simulation of age vectors) should be stored.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}
}
\value{
Returns an S3 object of class "bremla". This includes output from all functions nested within the bremla function. Including fitted marginals and summary statistics, simulated chronologies, time spent on each step.
}
\description{
Fits a regression model to the data and produces chronology simulations.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])


lowerints = which.index(event_intervals$depth_int_lower.m, depth[2:length(depth)])
upperints = which.index(event_intervals$depth_int_upper.m, depth[2:length(depth)])

eventnumber=13 #number between 1 and 29. specifies which transition to consider
depth.reference = event_intervals$NGRIP_depth_m[eventnumber]
age.reference = event_intervals$GICC_age.yb2k[eventnumber]
interval = lowerints[eventnumber]:upperints[eventnumber]

object = bremla(age,depth,proxy,events=eventdepths,nsims=10000,
  synchronization = list(locations=c(11050,12050,13050,22050,42050),
                          locations.type="age",method="adolphi",
                          samples=NULL),
  event.estimation = list(interval=interval,t1.sims=50000,rampsims=50000,label="GI-11",
                          depth.reference=event_intervals$NGRIP_depth_m[eventnumber],
                          age.reference=event_intervals$NGRIP_agee_m[eventnumber]),
  bias = list(bias.model="uniform",biasparams=cbind( c(1,1),c(0.98,1.02),c(0.96,1.04) ),
              store.samples=FALSE),
  print.progress=TRUE
  )
  summary(object)
  plot(object)
}



}
\seealso{
\code{\link{bremla_modelfitter},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
