% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla.R
\name{bremla}
\alias{bremla}
\title{Bayesian regression modeling of layered archives}
\usage{
bremla(
  age,
  depth,
  proxy,
  events = NULL,
  nsims = 10000,
  eventmeasure = "depth",
  reg.model = list(const = FALSE, depth1 = FALSE, depth2 = TRUE, proxy = TRUE, psi0 =
    TRUE, psi1 = TRUE),
  noise = "ar1",
  method = "inla",
  CI.type = "quantiles",
  DO.estimation = NULL,
  bias = NULL,
  store.everything = FALSE,
  print.progress = FALSE
)
}
\arguments{
\item{age}{Vector of observed ages y_0,...,y_n}

\item{depth}{Vector of depths z_0,...,z_n corresponding to \code{age}}

\item{proxy}{Vector of observed proxy (e.g. del-18O) at depths \code{depth}}

\item{events}{Vector describing locations (in dimension specified by \code{eventmeasure}) of climate transitions for use in linear regression model}

\item{nsims}{Number of chronologies to be simulated}

\item{eventmeasure}{Character describing in which dimension the climate events are located ("age", "depth", "index")}

\item{reg.model}{List of booleans specifies which effects to be included in the linear regression model: intercept (\code{const}), linear response to depth (\code{depth1}), quadratic response to depth (\code{depth2}), linear response to proxy (\code{proxy}), individual intercepts for each climate period (\code{psi0}), individual linear responses to depth for each climate period (\code{psi1})}

\item{noise}{Character specifying the noise model: independent identically distributed (\code{iid}), first order autoregressive (\code{ar1}) or second order autoregressive (\code{ar2})}

\item{method}{Character specifying how the model is fitted. Currently only least squares (\code{ls}) and INLA (\code{inla}) are supported, and least squares is always run.}

\item{CI.type}{Character describing which type of uncertainty intervals to be used. \code{CI.type="quantiles"} is computed from the quantiles, \code{CI.type="hpd"} computes highest posterior density intervals (takes a little longer to compute). If distribution is unimodal and symmetric these are equivalent.}

\item{DO.estimation}{List specifying the settings for the linear ramp fit and subsequent age simulation of a specified DO-event. See example.}

\item{bias}{List specifying the settings for the simulation of chronologies subjected to unknown counting bias. See example for implementation.}

\item{store.everything}{Boolean describing whether optional information (e.g. simulation of age vectors) should be stored.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}
}
\value{
Returns an S3 object of class "bremla". This includes output from all functions nested within the bremla function. Including fitted marginals and summary statistics, simulated chronologies, time spent on each step.
}
\description{
Fits a regression model to the data and produces chronology simulations.
}
\examples{

}
\seealso{
\code{\link{bremla_modelfitter},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
