% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_chronology_simulation.R
\name{bremla_chronology_simulation}
\alias{bremla_chronology_simulation}
\title{Bremla chronology simulation}
\usage{
bremla_chronology_simulation(object, control.sim, print.progress = FALSE)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{bremla_modelfitter}}}

\item{control.sim}{List object containing specifications for simulation procedure and
what is to be computed. See \code{\link{control.sim.default}} for details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}
}
\value{
Returns the same \code{object} list from the input, but appends simulated
chronologies, mean vectors (if \code{store.means=TRUE}) and further information, summary and statistics in
\code{object\$simulation}.
}
\description{
Simulates chronologies based on the fitted regression model.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])

nsims = 5000
object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=nsims)
object = bremla_modelfitter(object)
object = bremla_chronology_simulation(object,nsims=nsims)
plot(object,plot.inlasims = list(nsims=30,legend=NULL,xrev=FALSE,label=NULL))
}


}
\seealso{
\code{\link{bremla_prepare},\link{bremla_modelfitter}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{simulation}
