% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_chronology_simulation.R
\name{bremla_chronology_simulation}
\alias{bremla_chronology_simulation}
\title{Bremla chronology simulation}
\usage{
bremla_chronology_simulation(
  object,
  nsims = 10000,
  method = "inla",
  store.means = FALSE,
  print.progress = FALSE
)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{bremla_modelfitter}}}

\item{nsims}{Integer denoting the number of chronologies to be sampled}

\item{method}{Character specifying which method of inference to be used. Currently only \code{inla} is supported}

\item{store.means}{Boolean. If \code{TRUE} then simulated mean vectors will be stored. Default is \code{FALSE} to save memory.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}
}
\value{
Returns the same \code{object} list from the input, but appends simulated chronologies, and mean vectors (if \code{store.means=TRUE})
}
\description{
Simulates chronologies based on the fitted regression model.
}
\examples{

}
\seealso{
\code{\link{bremla_prepare},\link{bremla_modelfitter}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{simulation}
