% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_modelfitter (uit-mac-1070's conflicted
%   copy 2022-07-28).R, R/bremla_modelfitter.R
\name{bremla_modelfitter}
\alias{bremla_modelfitter}
\title{Bremla model fitting}
\usage{
bremla_modelfitter(object, control.fit, print.progress = FALSE)

bremla_modelfitter(object, control.fit, print.progress = FALSE)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{bremla_prepare}}}

\item{control.fit}{List containing specifications for fitting procedure. See
\code{\link{control.fit.default}} for details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}
}
\value{
Returns the same \code{object} list from the input, but appends information
from the fitting procedure in \code{object\$fitting}. Including fitted values, residuals, posterior distributions
and summary statistics for both fixed and random effects.

Returns the same \code{object} list from the input, but appends information
from the fitting procedure in \code{object\$fitting}. Including fitted values, residuals, posterior distributions
and summary statistics for both fixed and random effects.
}
\description{
Fits the bremla linear regression model to observations.

Fits the bremla linear regression model to observations.
}
\examples{
\donttest{
require(stats)
set.seed(1)
n <- 1000
phi <- 0.8
sigma <- 1.2
a_lintrend <- 0.3; a_proxy = 0.8
dy_noise <- as.numeric(arima.sim(model=list(ar=c(phi)),n=n,sd=sqrt(1-phi^2)))
lintrend <- seq(from=10,to=15,length.out=n)

proxy <- as.numeric(arima.sim(model=list(ar=c(0.9)),n=n,sd=sqrt(1-0.9^2)))
dy <- a_lintrend*lintrend + a_proxy*proxy + sigma*dy_noise

y0 = 11700;z0=1200
age = y0+cumsum(dy)
depth = 1200 + 1:n*0.05


formula = dy~-1+depth2 + proxy
data = data.frame(age=age,dy=dy,proxy=proxy,depth=depth,depth2=depth^2)
data = rbind(c(y0,NA,NA,z0,NA),data) #First row is only used to extract y0 and z0.

events=list(locations=c(1210,1220,1240))
control.fit = list(ncores=2,noise="ar1")
object = bremla_prepare(formula,data,nsims=5000,reference.label="simulated timescale",
                        events = events,
                        control.fit=control.fit)
object = bremla_modelfitter(object, print.progress=TRUE)
summary(object)
plot(object)
}
\donttest{
require(stats)
set.seed(1)
n <- 1000
phi <- 0.8
sigma <- 1.2
a_lintrend <- 0.3; a_proxy = 0.8
dy_noise <- as.numeric(arima.sim(model=list(ar=c(phi)),n=n,sd=sqrt(1-phi^2)))
lintrend <- seq(from=10,to=15,length.out=n)

proxy <- as.numeric(arima.sim(model=list(ar=c(0.9)),n=n,sd=sqrt(1-0.9^2)))
dy <- a_lintrend*lintrend + a_proxy*proxy + sigma*dy_noise

y0 = 11700;z0=1200
age = y0+cumsum(dy)
depth = 1200 + 1:n*0.05


formula = dy~-1+depth2 + proxy
data = data.frame(age=age,dy=dy,proxy=proxy,depth=depth,depth2=depth^2)
data = rbind(c(y0,NA,NA,z0,NA),data) #First row is only used to extract y0 and z0.

events=list(locations=c(1210,1220,1240))
control.fit = list(ncores=2,noise="ar1")
object = bremla_prepare(formula,data,nsims=5000,reference.label="simulated timescale",
                        events = events,
                        control.fit=control.fit)
object = bremla_modelfitter(object, print.progress=TRUE)
summary(object)
plot(object)
}
}
\seealso{
\code{\link{bremla_prepare},\link{bremla_chronology_simulation}}

\code{\link{bremla_prepare},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{fitting}
