% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_prepare.R
\name{bremla_prepare}
\alias{bremla_prepare}
\title{Bremla preparation function}
\usage{
bremla_prepare(
  age,
  depth,
  proxy,
  events = NULL,
  nsims = 10000,
  eventmeasure = "depth",
  reg.model = list(const = FALSE, depth1 = FALSE, depth2 = TRUE, proxy = TRUE, psi0 =
    TRUE, psi1 = TRUE),
  noise = "ar1",
  method = "inla",
  reference.label = NULL
)
}
\arguments{
\item{age}{Vector of observed ages y_0,...,y_n}

\item{depth}{Vector of depths z_0,...,z_n corresponding to \code{age}}

\item{proxy}{Vector of observed proxy (e.g. del-18O) at depths \code{depth}}

\item{events}{Vector describing locations (in dimension specified by \code{eventmeasure}) of climate transitions for use in linear regression model}

\item{nsims}{Number of chronologies to be simulated}

\item{eventmeasure}{Character describing in which dimension the climate events are located ("age", "depth", "index")}

\item{reg.model}{List of booleans specifies which effects to be included in the linear regression model: intercept (\code{const}), linear response to depth (\code{depth1}), quadratic response to depth (\code{depth2}), linear response to proxy (\code{proxy}), individual intercepts for each climate period (\code{psi0}), individual linear responses to depth for each climate period (\code{psi1})}

\item{noise}{Character specifying the noise model: independent identically distributed (\code{iid}), first order autoregressive (\code{ar1}) or second order autoregressive (\code{ar2})}

\item{method}{Character specifying how the model is fitted. Currently only least squares (\code{ls}) and INLA (\code{inla}) are supported, and least squares is always run.}

\item{reference.label}{String denoting the label used for age data/reference}
}
\value{
Returns a list containing data in internal formatting (\code{data}), linear regression formula (\code{ls.formula}), inla formula (\code{formula}), input settings and indices corresponding to climate transitions (\code{.args}) and z_0, y_0 and x_0 (\code{preceeding})
}
\description{
Prepares input and formats data for bremla and functions therein.
}
\examples{

}
\seealso{
\code{\link{bremla},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{preparation}
