% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{bremla_simulationsummarizer}
\alias{bremla_simulationsummarizer}
\title{Simulation-summarizer}
\usage{
bremla_simulationsummarizer(
  object,
  CI.type = "hpd",
  sync = TRUE,
  print.progress = FALSE
)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{bremla_chronology_simulation}}}

\item{CI.type}{character describing which uncertainty intervals should be used (\code{"quantiles" or "hpd"}). Highest posterior density is used by default}

\item{sync}{boolean. Set \code{TRUE} if summary statistics should be computed for syncrhonized samples (if any). \code{FALSE} for unsynchronized.}

\item{print.progress}{Boolean describing whether or not progress should be printed on screen.}
}
\value{
Returns the \code{object} list from the input and appends additional summary statistics: posterior marginal mean and uncertainty interval-
}
\description{
Computes posterior marginal mean and uncertainty intervals from simulations.
}
\seealso{
\code{\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{indices}
