% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_synchronized_simulation.R
\name{bremla_synchronized_simulation}
\alias{bremla_synchronized_simulation}
\title{Simulate synchronized chronologies}
\usage{
bremla_synchronized_simulation(object, control.sim, print.progress = FALSE)
}
\arguments{
\item{object}{Output of function \code{\link{tiepointsimmer}}.}

\item{control.sim}{List containing specifications for the simulation procedure,
including the number of samples to be generated and whether or not the chronologies
should be synchronized (for this function this should be set to \code{TRUE}).
See \code{\link{control.sim.default}} for details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen.}
}
\value{
Returns the \code{object} list from the input and appends a list which
includes all synchronized simulations, summary statistics and other information
computed or related to the simulation procedure in \code{object\$simulation}.
}
\description{
Produces simulations from the synchronized time scale given tie-point samples.
}
\examples{
\donttest{
require(stats)
n <- 1000
phi <- 0.8
sigma <- 1.2
a_lintrend <- 0.3; a_proxy = 0.8
dy_noise <- as.numeric(arima.sim(model=list(ar=c(phi)),n=n,sd=sqrt(1-phi^2)))
lintrend <- seq(from=10,to=15,length.out=n)

proxy <- as.numeric(arima.sim(model=list(ar=c(0.9)),n=n,sd=sqrt(1-0.9^2)))
dy <- a_lintrend*lintrend + a_proxy*proxy + sigma*dy_noise

y0 = 11700;z0=1200
age = y0+cumsum(dy)
depth = 1200 + 1:n*0.05


formula = dy~-1+depth2 + proxy
data = data.frame(age=age,dy=dy,proxy=proxy,depth=depth,depth2=depth^2)
data = rbind(c(y0,NA,NA,z0,NA),data) #First row is only used to extract y0 and z0.

events=list(locations=c(1210,1220,1240))
control.fit = list(ncores=2,noise="ar1")
synchronization=list(method="gauss")
control.sim=list(synchronized=TRUE,
                 summary=list(compute=TRUE))

object = bremla_prepare(formula,data,nsims=5000,reference.label="simulated timescale",
                        events = events,
                        synchronization=synchronization,
                        control.fit=control.fit,
                        control.sim=control.sim)
object = bremla_modelfitter(object)
object = tiepointsimmer(object)
object = bremla_synchronized_simulation(object, print.progress=TRUE)
summary(object)
plot(object)
}
}
\seealso{
\code{\link{bremla_chronology_simulation}} \code{\link{bremla}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{simulation}
\keyword{synchronization}
\keyword{tiepoint}
