% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bremla_synchronized_simulation.R
\name{bremla_synchronized_simulation}
\alias{bremla_synchronized_simulation}
\title{Simulate synchronized chronologies}
\usage{
bremla_synchronized_simulation(object, control.sim, print.progress = FALSE)
}
\arguments{
\item{object}{Output of function \code{\link{tiepointsimmer}}.}

\item{control.sim}{List containing specifications for the simulation procedure,
including the number of samples to be generated and whether or not the chronologies
should be synchronized (for this function this should be set to \code{TRUE}).
See \code{\link{control.sim.default}} for details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen.}
}
\value{
Returns the \code{object} list from the input and appends a list which
includes all synchronized simulations, summary statistics and other information
computed or related to the simulation procedure in \code{object\$simulation}.
}
\description{
Produces simulations from the synchronized time scale given tie-point samples.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])

nsims = 5000
object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=nsims)
object = bremla_modelfitter(object)
object = bremla_chronology_simulation(object,nsims=nsims)
object = tiepointsimmer(object,nsims=nsims,method="adolphi")
object = bremla_synchronized_simulation(object,nsims=nsims)
object = bremla_simulationsummarizer(object,CI.type="hpd",sync=TRUE)
plot(object)
}


}
\seealso{
\code{\link{bremla_chronology_simulation}} \code{\link{bremla}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{simulation}
\keyword{synchronization}
\keyword{tiepoint}
