% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_arguments.R
\name{control.fit.default}
\alias{control.fit.default}
\title{Default variables in control.fit}
\usage{
control.fit.default()
}
\value{
Returns a list including default values for all variables in \code{control.fit}.
}
\description{
Sets the default variables in the list \code{control.fit} used to specify the
fitting procedure. The list contains the following arguments:
\itemize{
  \item{\code{noise} }{Character which noise process is assumed for the residuals.
  \code{"iid"}, \code{"ar1"} and \code{"ar2"} are currently. However, if synchronization is to
  be performed only \code{"ar1"} is currently supported. Default input is \code{"ar1"}.}
  \item{\code{method} }{Character describing which method should be used to fit the process.
  Only \code{inla} is currently supported, but more will hopefully come in the future.}
  \item{\code{verbose} }{Boolean telling whether or not the \code{inla}-program should run in
  verbose mode. Default \code{FALSE}.}
  \item{\code{ncores} }{Integer used to set the number of cores to be used in the \code{inla} fitting
  procedure. Default value is \code{1}.}
  \item{\code{transform} }{Character indicating which transformation should be applied to
  the observations. Can be \code{"log"} or \code{"identity"}. If synchronization is
  intended then only \code{"identity"} is currently supported, which is also the
  default value here.}
}
}
\seealso{
\code{\link{bremla},\link{bremla_modelfitter},\link{set.options}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{default}
\keyword{fitting}
