% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_depth_to_age.R
\name{events_depth_to_age}
\alias{events_depth_to_age}
\title{Complete dating uncertainty of abrupt warming transitions}
\usage{
events_depth_to_age(object, control.transition_dating, print.progress = FALSE)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{linrampfitter}}}

\item{control.transition_dating}{List containing speci}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen}
}
\value{
Returns the same \code{object} list from the input, but appends the simulated transition onset ages along with summary statistics.
}
\description{
Combines linear onset posterior from ramp model fit and simulated chronologies from
the Bayesian regression modeling to estimate complete dating uncertainty of the onset
of abrupt warming transitions events using Monte Carlo simulation.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])

nsims = 5000
object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=nsims)
object = bremla_modelfitter(object)
object = bremla_chronology_simulation(object,nsims=nsims)
object = tiepointsimmer(object,nsims=nsims,method="adolphi")
object = bremla_synchronized_simulation(object,nsims=nsims)
object = bremla_simulationsummarizer(object,CI.type="hpd",sync=TRUE)

lowerints = which.index(event_intervals$depth_int_lower.m, depth[2:length(depth)])
upperints = which.index(event_intervals$depth_int_upper.m, depth[2:length(depth)])

eventnumber=13 #number between 1 and 29. specifies which transition to consider
depth.reference = event_intervals$NGRIP_depth_m[eventnumber]
age.reference = event_intervals$GICC_age.yb2k[eventnumber]
interval = lowerints[eventnumber]:upperints[eventnumber]

object = linrampfitter(object,interval,label="GI-11",depth.reference=depth.reference)
object = events_depth_to_age(object,nsims=nsims,age.reference=age.reference)
plot(object,plot.proxydata=list(age=TRUE,depth=FALSE,xrev=FALSE,label=NULL),
    plot.ls = NULL,
    plot.inla.posterior = NULL,
    plot.inlasims = NULL,
    plot.syncsims = NULL,
    plot.tiepoints = NULL,
    plot.bias = NULL,
    plot.linramp = list(depth.reference=NULL,show.t0=TRUE,show.t1=TRUE,xrev=TRUE,label=NULL))
}

}
\seealso{
\code{\link{bremla_chronology_simulation},\link{linrampfitter}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{dating}
\keyword{transition}
