% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linrampfitter.R
\name{linrampfitter}
\alias{linrampfitter}
\title{Linear ramp fit}
\usage{
linrampfitter(object, control.linramp, print.progress = FALSE)
}
\arguments{
\item{object}{List object which is the output of function \code{\link{bremla_chronology_simulation}}}

\item{control.linramp}{Includes the data and specifications for the linear ramp
model fitting procedure. Must include \code{control.linramp\$proxy} and \code{control.linramp\$interval}
(preferably also \code{control.linramp\$interval.unit}),
the rest can be set to default values. See \code{\link{control.linramp.default}} for details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen.}
}
\value{
Returns the same \code{object} list from the input, but appends another list of results from the linear ramp fit, including posterior marginal distributions of the hyperparameters, summary statistics, and simulations of linear ramp and transition end point (if any).
}
\description{
Fits the linear ramp described by Erhardt et al. (2019) to proxy values using INLA.
}
\examples{

\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])

object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=0)

lowerints = which.index(event_intervals$depth_int_lower.m, depth[2:length(depth)])
upperints = which.index(event_intervals$depth_int_upper.m, depth[2:length(depth)])

eventnumber=13 #number between 1 and 29. specifies which transition to consider
depth.reference = event_intervals$NGRIP_depth_m[eventnumber]
interval = lowerints[eventnumber]:upperints[eventnumber]

object = linrampfitter(object,interval,label="GI-11",depth.reference=depth.reference)

plot(object,plot.proxydata=list(age=TRUE,depth=FALSE,xrev=FALSE,label=NULL),
    plot.ls = NULL,
    plot.inla.posterior = NULL,
    plot.inlasims = NULL,
    plot.syncsims = NULL,
    plot.tiepoints = NULL,
    plot.bias = NULL,
    plot.linramp = list(depth.reference=NULL,show.t0=TRUE,show.t1=TRUE,xrev=TRUE,label=NULL))

}

}
\references{
Erhardt et al. (2019)
}
\seealso{
\code{\link{bremla_chronology_simulation},\link{events_depth_to_age}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{linear_ramp}
