% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bremla.R
\name{print.bremla}
\alias{print.bremla}
\title{Print bremla model}
\usage{
print.bremla(x, digits = 4L, ...)
}
\arguments{
\item{x}{\code{bremla} S3 class. Output of \code{\link{bremla}} function}

\item{digits}{Number of digits displayed.}

\item{...}{Other arguments}
}
\description{
Prints essentials from bremla S3 class.
}
\examples{

}
\seealso{
\code{\link{bremla},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{sumary}
