% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bremla.R
\name{print.summary.bremla}
\alias{print.summary.bremla}
\title{Print summary.bremla class}
\usage{
\method{print}{summary.bremla}(x, digits = 4L, ...)
}
\arguments{
\item{x}{\code{summary.bremla} S3 class. Output of \code{\link{summary.bremla}} function}

\item{digits}{Number of digits displayed.}

\item{...}{Other arguments}
}
\description{
Prints summary.bremla S3 class.
}
\examples{

}
\seealso{
\code{\link{bremla},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{summary.print}
