% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{Qsimmer}
\alias{Qsimmer}
\title{Efficient Gaussian simulation
Simulate from multivariate Gaussian distribution efficiently by using sparse precision matrix}
\usage{
Qsimmer(nsims, Q, muvek)
}
\arguments{
\item{nsims}{The number of samples to be produced.}

\item{Q}{The precision matrix, should be sparse.}

\item{muvek}{The mean vector.}
}
\value{
Returns matrix with nsims samples of length equal to \code{nrow(Q)}.
}
\description{
Efficient Gaussian simulation
Simulate from multivariate Gaussian distribution efficiently by using sparse precision matrix
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{GMRF}
\keyword{simulation}
\keyword{sparse}
