% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{Qymaker}
\alias{Qymaker}
\title{Precision matrix of cumulative Gaussian process
Produces the sparse precision matrix of a cumulative Gaussian process}
\usage{
Qymaker(Qx)
}
\arguments{
\item{Qx}{The precision matrix of the Gaussian process to be transformed.}
}
\value{
Returns sparse precision matrix of dimension \code{n x n}.
}
\description{
Precision matrix of cumulative Gaussian process
Produces the sparse precision matrix of a cumulative Gaussian process
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{GMRF}
\keyword{matrix}
\keyword{precision}
\keyword{sparse}
