% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bremla.R
\name{summary.bremla}
\alias{summary.bremla}
\title{Summary bremla model}
\usage{
\method{summary}{bremla}(object, digits = 4L, ...)
}
\arguments{
\item{object}{\code{bremla} S3 class. Output of \code{\link{bremla}} function}

\item{digits}{Number of digits displayed.}

\item{...}{Other arguments}
}
\description{
Summarizes results from bremla S3 class.
}
\examples{

}
\seealso{
\code{\link{bremla},\link{bremla_chronology_simulation}}
}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{bremla}
\keyword{sumary}
