% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfulfunctions.R
\name{tiepointsimmer}
\alias{tiepointsimmer}
\title{Tie-point simulation}
\usage{
tiepointsimmer(object, synchronization, print.progress = FALSE, ...)
}
\arguments{
\item{object}{Bremla list object to store samples in}

\item{synchronization}{List containing specifications for generating tie-points.
Must include \code{synchronization\$locations}. Includes different methods for
producing tie-points specified by \code{synchronization\$method}, and can also
import pre-computed samples in \code{synchronization\$samples}. See
\code{\link{synchronization.default}} for more details.}

\item{print.progress}{Boolean. If \code{TRUE} progress will be printed to screen.}

\item{...}{Further arguments to be passed down to the selected tie-point simulation
function.}
}
\value{
returns the same \code{object} from input but appends tie-point samples
and information in \code{object\$tie_points}.
}
\description{
Wrapper function which calls specified functions to produce samples from tie-point distributions and places them in \code{bremla} object.
}
\examples{
\donttest{
data("event_intervals")
data("events_rasmussen")
data("NGRIP_5cm")

age = NGRIP_5cm$age
depth = NGRIP_5cm$depth
d18O = NGRIP_5cm$d18O
proxy=d18O

eventdepths = events_rasmussen$depth
eventindexes = c(1,which.index(eventdepths, depth[2:length(depth)]) )
eventindexes = unique(eventindexes[!is.na(eventindexes)])

object = bremla_prepare(age,depth,proxy,events=eventdepths,nsims=0)
object = tiepointsimmer(object,nsims=10000,method="adolphi")
hist(object$tie_points$samples[,2],col="orange",freq=0,breaks=40,
    main="Tie-point 3",xlab="Age (yb2k)")
}


}
\author{
Eirik Myrvoll-Nilsen, \email{eirikmn91@gmail.com}
}
\keyword{sample}
\keyword{tiepoint}
